+++ src/reducers/diner.js
@@ -1,0 +1,14 @@
+const initialState = {
+  username: "",
+  email: "",
+  password: "",
+  currentLocation: "",
+  favoriteTrucks: [],
+};
+
+export const diner = (state = initialState, action) => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
