--- src/reducers/dinerReducer.js
@@ -1,16 +1,0 @@
-const initialState = {
-  diner: {
-    username: "",
-    email: "",
-    password: "",
-    currentLocation: "",
-    favoriteTrucks: [],
-  },
-};
-
-export const dinerReducer = (state = initialState, action) => {
-  switch (action.type) {
-    default:
-      return state;
-  }
-};
